pyproject.toml
[project]
name = "quarto-mcp-server"
version = "0.1.0"
description = "MCP server for converting Quarto Markdown to PowerPoint and other formats"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "CC-BY-4.0"}
authors = [
    {name = "Quarto MCP Server Contributors"}
]
keywords = ["mcp", "quarto", "markdown", "powerpoint", "presentation", "conversion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
quarto-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
